medallion-weather/pyproject.toml
[tool.poetry]
name = "medallion_weather"
description = "This project demonstrates implementation of Medallion Architecture proposed by Databricks using Dagster."
version = "0.1.2"
authors = ["George T. C. Lai <tsungchih.hd@gmail.com>"]
packages = [
    { include = "medallion_weather" },
]

[tool.poetry.dependencies]
python = ">=3.9, <3.10"
requests = "^2.28.0"
python-dotenv = "^0.21.0"
google-cloud-storage = "^2.3.0"
modin = {extras = ["all"], version = "0.17.0"}
orjson = "3.8.0"
pandas = "^1.4.0"
pyarrow = "^8.0.0"
dagster = "1.2.1"
dagster-graphql = "1.2.1"
dagster-postgres = "0.18.1"
dagster-k8s = "0.18.1"
dagster-gcp = "0.18.1"
dagster-pandera = "0.18.1"

[tool.poetry.group.dev.dependencies]
pytest = "^5.2"
pylama = "^8.4.1"
coverage = "^6.4.1"
dagit = "^1.2.1"
icecream = "^2.1.3"

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
